api_module/requirements.txt
flask==2.3.3
anthropic==0.16.0
python-dotenv==1.0.0
gunicorn==21.2.0 